(1209) feat: N+1 문제 측정을 위한 테스트 코드 작성  - PostServiceNPlusOneTest 작성  - 게시글 목록 조회 시 N+1 문제 확인 (21개 쿼리 발생)  - 테스트용 application.properties SQL 로깅 활성화  - 실행 시간 및 예상 쿼리 수 측정

+++ src/test/java/com/cho/board/integration/PostIntegrationTest.java
@@ -174,7 +174,7 @@ class PostIntegrationTest {
             PostFixture.createPost("조회수 테스트", "내용", testUser, testCategory)
         );
 
-        int initialViewCount = post.getViewCount();
+        Long initialViewCount = post.getViewCount();
 
         // when & then
         mockMvc.perform(get("/posts/{id}", post.getId()))
